Move pycodestyle rules into individual files
