+++ src/rules/pycodestyle/rules/line_too_long.rs
@@ -1,0 +1,30 @@
+use rustpython_ast::Location;
+
+use crate::ast::types::Range;
+use crate::registry::Diagnostic;
+use crate::rules::pycodestyle::rules::is_overlong;
+use crate::settings::Settings;
+use crate::violations;
+
+/// E501
+pub fn line_too_long(lineno: usize, line: &str, settings: &Settings) -> Option<Diagnostic> {
+    let line_length = line.chars().count();
+    let limit = settings.line_length;
+    if is_overlong(
+        line,
+        line_length,
+        limit,
+        settings.pycodestyle.ignore_overlong_task_comments,
+        &settings.task_tags,
+    ) {
+        Some(Diagnostic::new(
+            violations::LineTooLong(line_length, limit),
+            Range::new(
+                Location::new(lineno + 1, limit),
+                Location::new(lineno + 1, line_length),
+            ),
+        ))
+    } else {
+        None
+    }
+}
